[AIRFLOW-2110] Code review
